requirements.txt
aniso8601==8.0.0
apipkg==1.4
appdirs==1.4.3
appnope==0.1.0
asn1crypto==0.24.0
astroid==2.4.2
atomicwrites==1.1.5
attrs==19.3.0
autopep8==1.5.5
beautifulsoup4==4.9.3
black==21.7b0
bs4==0.0.1
certifi==2018.1.18
cffi==1.14.0
chardet==3.0.4
coverage==5.4
entrypoints==0.3
flake8==3.8.4
Flask==2.0.1
Flask-Compress==1.4.0
Flask-Cors==3.0.10
Flask-JWT-Extended==3.24.1
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
flask-talisman==0.8.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.9
iniconfig==1.1.1
isort==5.7.0
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
msgpack==1.0.2
packaging==20.3
pluggy==0.13.1
psycopg2==2.8.4
py==1.10.0
pycodestyle==2.6.0
pycparser==2.18
pycryptodome==3.9.7
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.6.2
PyNaCl==1.3.0
pyparsing==2.4.6
pyrsistent==0.15.7
pytest==6.2.2
pytest-cov==2.11.1
python-http-client==3.2.6
pytz==2019.3
redis==3.5.3
regex==2020.2.20
requests==2.25.1
sendgrid==6.1.3
six==1.15.0
soupsieve==2.0
SQLAlchemy==1.4.27
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.26.6
wcwidth==0.1.8
Werkzeug==2.0.1
wrapt==1.11.2
WTForms==2.2.1
zxcvbn==4.4.28
